rename workflowCloneModal folder name
